backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
Pillow==10.1.0
transformers==4.36.2
torch==2.1.2
numpy<2
python-dotenv==1.0.0
